moved ParseHistoryBeamHeader function from Casacore to String
