+++ test/Util/TestUtilString.cc
@@ -7,6 +7,7 @@
 #include <gtest/gtest.h>
 #include <string>
 #include <vector>
+#include "Util/Casacore.h"
 #include "Util/String.h"
 
 TEST(StringUtilTest, SplitString_Basic) {
@@ -107,4 +108,30 @@ TEST(StringUtilTest, HasSuffixCaseInsensitive) {
     EXPECT_FALSE(HasSuffix("test.fits", ".xml"));
     EXPECT_TRUE(HasSuffix("test.fits.gz", ".fits.gz"));
     EXPECT_FALSE(HasSuffix("test.fits.gz", ".fits"));
+}
+
+TEST(ParseHistoryBeamHeaderTest, ValidHistoryBeamFormat) {
+    std::string header = "HISTORY RESTOR Beam = 2.000E+00 x 1.800E+00 arcsec, pa = 8.000E+01 degrees";
+    std::string bmaj, bmin, bpa;
+
+    EXPECT_TRUE(ParseHistoryBeamHeader(header, bmaj, bmin, bpa));
+    EXPECT_EQ(bmaj, "2.000E+00arcsec");
+    EXPECT_EQ(bmin, "1.800E+00arcsec");
+    EXPECT_EQ(bpa, "8.000E+01deg");
+}
+
+TEST(ParseHistoryBeamHeaderTest, InvalidHistoryBeamFormat) {
+    std::string header = "Invalid format";
+    std::string bmaj, bmin, bpa;
+
+    EXPECT_FALSE(ParseHistoryBeamHeader(header, bmaj, bmin, bpa));
+}
+
+TEST(ParseHistoryBeamHeaderTest, ValidGaussianBeam) {
+    casacore::GaussianBeam beam(casacore::Quantity(1.5, "arcsec"), casacore::Quantity(1.2, "arcsec"), casacore::Quantity(45, "deg"));
+
+    std::string formatted = FormatBeam(beam);
+    EXPECT_NE(formatted.find("major: 1.500000 arcsec"), std::string::npos);
+    EXPECT_NE(formatted.find("minor: 1.200000 arcsec"), std::string::npos);
+    EXPECT_NE(formatted.find("pa: 45.000000 deg"), std::string::npos);
 }
